chore(chat-custom-width/cleanup): Remove unused files
- Remove unused JavaScript files to reduce repository size and improve maintainability

Changes summary:
- Removed unused JavaScript files, reducing clutter and improving the overall maintainability of the codebase.
