+++ src/js/app/components/renderSwitch.js
@@ -1,22 +1,31 @@
 import { icon_lock } from './icons'
 
 function renderSwitchOption({ inputId, isChecked = false, icon, textTitle, textSubtitle }) {
+	const sanitizedInputId = sanitizeString(inputId)
+	const sanitizedTextTitle = sanitizeString(textTitle)
+	const sanitizedTextSubtitle = sanitizeString(textSubtitle)
+
 	return `
-        <label class="gpth-switch" for="${inputId}">
-            <div class="gpth-switch__icon">
-                ${icon}
-            </div>
+    <label class="gpth-switch" for="${sanitizedInputId}">
+      <div class="gpth-switch__icon" aria-hidden="true">
+        ${icon}
+      </div>
 
-            <div class="gpth-switch__text">
-                <div class="title">${textTitle}</div>
-                <div class="subtitle">${textSubtitle}</div>
-            </div>
+      <div class="gpth-switch__text">
+        <div class="title">${sanitizedTextTitle}</div>
+        <div class="subtitle">${sanitizedTextSubtitle}</div>
+      </div>
 
-            <div class="gpth-switch__checkbox">
-                <input type="checkbox" id="${inputId}" ${isChecked ? 'checked' : ''}>
-                <span class="slider"></span>
-            </div>
-        </label>`
+      <div class="gpth-switch__checkbox">
+        <input 
+          type="checkbox" 
+          id="${sanitizedInputId}" 
+          ${isChecked ? 'checked' : ''}
+          aria-labelledby="${sanitizedInputId}-label"
+        >
+        <span class="slider" aria-hidden="true"></span>
+      </div>
+    </label>`
 }
 
 function renderSmallCardOption({
@@ -30,28 +39,65 @@ function renderSmallCardOption({
 	unit = '%',
 	isLocked = false,
 }) {
+	const sanitizedName = sanitizeString(name)
+	const sanitizedInputId = sanitizeString(inputId)
+	const sanitizedInputValue = sanitizeString(inputValue)
+	const sanitizedInputPlaceholder = sanitizeString(inputPlaceholder)
+
 	const lockIcon = isLocked ? icon_lock : ''
 
-	// <label for="${inputId}" class="grid justify-center content-space-between p-2 gap-4 h-full w-full rounded-xl">
-
 	return `
-        <div class="card card--range ${isLocked ? 'is-locked' : ''}" data-gpth-err="${min}${unit} &hArr; ${max}${unit}">
-            <label for="${inputId}" class="flex flex-col justify-center p-2 gap-4 h-full w-full rounded-xl">
-                <div class="flex items-center gap-2 w-full">
-                    <div class="card__output flex-none h-10 w-10 font-semibold rounded-full grid items-center justify-center" id="range-output-${inputId}">
-                        100
-                    </div>
-                    <div class="card__unitname-wrapper grid">
-                        <p class="card__unit rounded-full flex items-center justify-center mb-2 font-semibold" id="unit-${inputId}">${unit}</p>
-                        <p class="card__name uppercase font-semibold w-full break-words">${name}</p>
-                    </div>
-                </div>
+    <div 
+      class="card card--range ${isLocked ? 'is-locked' : ''}" 
+      data-gpth-err="${min}${unit} &hArr; ${max}${unit}"
+    >
+      <label 
+        for="${sanitizedInputId}" 
+        class="flex flex-col justify-center p-2 gap-4 h-full w-full rounded-xl"
+      >
+        <div class="flex items-center gap-2 w-full">
+          <div 
+            class="card__output flex-none h-10 w-10 font-semibold rounded-full grid items-center justify-center" 
+            id="range-output-${sanitizedInputId}"
+            aria-live="polite"
+          >
+            ${sanitizedInputValue}
+          </div>
+          <div class="card__unitname-wrapper grid">
+            <p 
+              class="card__unit rounded-full flex items-center justify-center mb-2 font-semibold" 
+              id="unit-${sanitizedInputId}"
+            >
+              ${unit}
+            </p>
+            <p class="card__name uppercase font-semibold w-full break-words">
+              ${sanitizedName}
+            </p>
+          </div>
+        </div>
 
-                <input type="${inputType}" id="${inputId}" value="${inputValue}" placeholder="${inputPlaceholder}" class="outline-none border-none" minlength="${min}" maxlength="${max}">
+        <input 
+          type="${inputType}" 
+          id="${sanitizedInputId}" 
+          value="${sanitizedInputValue}" 
+          placeholder="${sanitizedInputPlaceholder}" 
+          class="outline-none border-none" 
+          min="${min}" 
+          max="${max}"
+          aria-valuemin="${min}"
+          aria-valuemax="${max}"
+          aria-valuenow="${sanitizedInputValue}"
+          ${isLocked ? 'disabled' : ''}
+        >
+      </label>
+      ${lockIcon}
+    </div>`
+}
 
-            </label>
-            ${lockIcon}
-        </div>`
+function sanitizeString(str) {
+	const div = document.createElement('div')
+	div.textContent = str
+	return div.innerHTML
 }
 
 export { renderSwitchOption, renderSmallCardOption }
