+++ src/js/content.js
@@ -1,7 +1,5 @@
 import './app/floatingBtn'
 import './app/mainFonts'
-// import { init as initAssets } from './app/mainAssets'
-// import { init as initAssets } from './app/mainAssetsResize'
-import { init as initAssets } from './app/mainAssetsRefactor'
+import { init as initAssets } from './app/mainAssets'
 
 initAssets()
